some cart views implemented

+++ cart/views.py
@@ -5,7 +5,7 @@ from rest_framework.permissions import IsAdminUser, IsAuthenticated
 from django.shortcuts import get_object_or_404
 
 from core.models import ShoppingCart, ShoppingCartItem, Book
-from .serializers import ShoppingCartSerializer
+from .serializers import ShoppingCartSerializer, ShoppingCartItemSerializer
 
 @api_view(["GET"])
 @permission_classes([IsAdminUser, IsAuthenticated])
@@ -28,17 +28,40 @@ def get_shopping_cart(request, pk):
     serializer = ShoppingCartSerializer(cart)
     return Response(serializer.data, status=status.HTTP_200_OK)
 
+
 @api_view(["POST"])
-def add_item_to_cart(request, book_id):
+def add_item_to_cart(request):
     """
     Add a book in the user cart
     """
 
-    book = get_object_or_404(Book, pk=book_id)
-    cart = get_object_or_404(ShoppingCart, user=request.user)
-    item, _ = ShoppingCartItem.objects.get_or_create(book=book)
-    cart.items.add(item)
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    try:
+        book_id = int(request.data.get("book_id", ''))
+        quantity = request.data.get("quantity", '')
+    except ValueError:
+        return Response({
+            "error": "The 'book_id' and 'quantity' must be intengers values."
+        }, status=status.HTTP_400_BAD_REQUEST)
+
+    books = Book.objects.all()
+    if not books.filter(pk=book_id).exists():
+        return Response({
+            "error": "The book '{0}' does not exists".format(book_id)
+        }, status=status.HTTP_404_NOT_FOUND)
+    book = books.get(pk=book_id)
+
+    cart = ShoppingCart.objects.get(user=request.user)
+
+    if cart.items.filter(book=book).exists():
+        existent_item = cart.items.get(book=book)
+        existent_item.quantity = quantity
+        existent_item.save()
+    else:
+        item = ShoppingCartItem.objects.create(book=book, quantity=quantity)
+        cart.items.add(item)
+
+    serializer = ShoppingCartSerializer(cart)
+    return Response(serializer.data, status=status.HTTP_200_OK)
 
 @api_view(["DELETE"])
 def remove_item_from_cart(request, item_id):
@@ -46,6 +69,55 @@ def remove_item_from_cart(request, item_id):
     Remove specified item from user cart
     """
 
-    item = get_object_or_404(ShoppingCartItem, pk=item_id)
+    cart = ShoppingCart.objects.get(user=request.user)
+    if not cart.items.filter(pk=item_id).exists():
+        return Response({
+            "error": "The item '{0}' is not in the cart.".format(item_id)
+        }, status=status.HTTP_404_NOT_FOUND)
+
+    item = cart.items.get(pk=item_id)
     item.delete()
-    return Response(status=status.HTTP_204_NO_CONTENT)
+    serializer = ShoppingCartSerializer(cart)
+    return Response(serializer.data, status=status.HTTP_200_OK)
+
+@api_view(["GET"])
+def get_user_cart_items(request):
+    """
+    Get all user shoppingcart items in database
+    """
+
+    cart = ShoppingCart.objects.get(user=request.user)
+    serializer = ShoppingCartItemSerializer(cart.items, many=True)
+    return Response(serializer.data, status=status.HTTP_200_OK)
+
+@api_view(["PUT"])
+def update_cart_item(request):
+    """
+    Update a shoppingcart item
+    """
+
+    try:
+        book_id = int(request.data.get("book_id", ''))
+        quantity = int(request.data.get("quantity", ''))
+    except ValueError:
+        return Response({
+            "error": "The 'book_id' and 'quantity' must be intengers values."
+        }, status=status.HTTP_400_BAD_REQUEST)
+
+    books = Book.objects.all()
+    if not books.filter(pk=book_id).exists():
+        return Response({
+            "error": "The book '{0}' does not exists".format(book_id)
+        }, status=status.HTTP_404_NOT_FOUND)
+    book = books.get(pk=book_id)
+    cart = ShoppingCart.objects.get(user=request.user)
+
+    item = cart.items.filter(book=book)
+    if not item.exists():
+        return Response({
+            "error": "This book is not in the cart."
+        }, status=status.HTTP_400_BAD_REQUEST)
+
+    item.update(quantity=quantity)
+    serializer = ShoppingCartSerializer(cart)
+    return Response(serializer.data, status=status.HTTP_200_OK)
